@@ -1,10 +1,17 @@
 {
     "name": "shouwyee",
     "version": "0.0.1",
-    "description": "Aoi.js, but coded using TypeScript.",
+    "description": "Aoi.js, but coded using TypeScript and better interpreter (maybe)",
     "type": "commonjs",
     "main": "dist/index.js",
-    "keywords": ["discord.js", "aoi.js", "shouwyee", "bdfd", "discord", "bot"],
+    "keywords": [
+        "discord.js",
+        "aoi.js",
+        "shouwyee",
+        "bdfd",
+        "discord",
+        "bot"
+    ],
     "scripts": {
         "format": "biome format --write src",
         "dev": "node test"
@@ -16,8 +23,8 @@
         "discord.js": "^14.18.0"
     },
     "devDependencies": {
-        "@biomejs/biome": "1.9.4",
-        "typescript": "^5.7.3"
+        "@biomejs/biome": "^1.9.4",
+        "typescript": "^5.8.2"
     },
     "repository": {
         "type": "git",
@@ -36,5 +43,10 @@
     "bugs": {
         "url": "https://github.com/tyowk/shouwyee/issues"
     },
-    "files": ["dist", "README.md", "LICENSE", "package.json"]
+    "files": [
+        "dist",
+        "README.md",
+        "LICENSE",
+        "package.json"
+    ]
 }
